Last work before giving to moheb

+++ WindowsFormsApp3/OfficePay.cs
@@ -13,6 +13,9 @@ namespace WindowsFormsApp3
 {
     public partial class OfficePay : Form
     {
+
+        public bool isUpdate = false;
+
         public OfficePay()
         {
             InitializeComponent();
@@ -26,29 +29,55 @@ namespace WindowsFormsApp3
                 return;
 
             }
-            
-            SQLiteConnection scn = new SQLiteConnection(@"data source = main.db");
-            scn.Open();
-            SQLiteCommand sq;
+            if (!isUpdate)
+            {
+                SQLiteConnection scn = new SQLiteConnection(@"data source = main.db");
+                scn.Open();
+                SQLiteCommand sq;
+                
 
-            sq = new SQLiteCommand("delete from pay where fileno = '" + fileno.Text + "' and payer = '" + payer.Text + "' and officevoucher = '1'", scn);
-            sq.ExecuteNonQuery();
+                sq = new SQLiteCommand(String.Format("insert into pay (date,payer,fileno,slipno,amount,received,remarks,officevoucher,title) values ('{0}','{1}','{2}','{3}','{4}','{5}','{6}','{7}','{8}')",
+                   dateTimePicker1.Text,
+                   payer.Text,
+                   fileno.Text,
+                   slipno.Text,
+                   amount.Text,
+                   receive.Text,
+                   remarks.Text,
+                   "1",
+                   titleBox.Text), scn);
+
+                sq.ExecuteNonQuery();
+
+                MessageBox.Show("Data Saved successfully");
+                Close();
+            }
+            else
+            {
+                SQLiteConnection scn = new SQLiteConnection(@"data source = main.db");
+                scn.Open();
+                SQLiteCommand sq;
+
+                sq = new SQLiteCommand("delete from pay where fileno = '" + fileno.Text + "' and payer = '" + payer.Text + "' and officevoucher = '1'", scn);
+                sq.ExecuteNonQuery();
 
 
-            sq = new SQLiteCommand(String.Format("insert into pay (date,payer,fileno,slipno,amount,received,remarks,officevoucher) values ('{0}','{1}','{2}','{3}','{4}','{5}','{6}','{7}')",
-               dateTimePicker1.Text,
-               payer.Text,
-               fileno.Text,
-               slipno.Text,
-               amount.Text,
-               receive.Text,
-               remarks.Text,
-               "1"), scn);
+                sq = new SQLiteCommand(String.Format("insert into pay (date,payer,fileno,slipno,amount,received,remarks,officevoucher,title) values ('{0}','{1}','{2}','{3}','{4}','{5}','{6}','{7}','{8}')",
+                   dateTimePicker1.Text,
+                   payer.Text,
+                   fileno.Text,
+                   slipno.Text,
+                   amount.Text,
+                   receive.Text,
+                   remarks.Text,
+                   "1",
+                   titleBox.Text), scn);
 
-            sq.ExecuteNonQuery();
+                sq.ExecuteNonQuery();
 
-            MessageBox.Show("Data Saved successfully");
-            Close();
+                MessageBox.Show("Data Saved successfully");
+                Close();
+            }
         }
 
         private void OfficePay_Load(object sender, EventArgs e)
